Add send transaction test
