+++ universal-login-web3/test/ULWeb3Provider.ts
@@ -1,14 +1,38 @@
 import {expect} from 'chai';
 import {utils, Wallet} from 'ethers';
 import {createFixtureLoader} from 'ethereum-waffle';
-import {waitExpect, waitUntil} from '@universal-login/commons';
-import {providerFixture} from './fixtures/provider';
+import {waitExpect} from '@universal-login/commons';
+import {oneWallet} from './fixtures/oneWalletFixture';
+import Web3 from 'web3';
+import {Provider} from 'web3/providers';
+import {ULWeb3Provider} from '../lib';
+import {AppProps} from '../lib/ui/App';
 
 const loadFixture = createFixtureLoader();
 
+function createProvider(provider: Provider, relaterUrl: string): [ULWeb3Provider, AppProps] {
+  let services: AppProps;
+  const ulProvider = new ULWeb3Provider(provider, relaterUrl, ['mylogin.eth'], (props: AppProps) => {
+    services = props;
+  });
+
+  return [ulProvider, services!];
+}
+
+
 describe('ULWeb3Provier', () => {
+  const setup = async () => {
+    const {relayer, provider, deployedWallet} = await loadFixture(oneWallet);
+    const [ulProvider, services] = createProvider(provider._web3Provider as any, relayer.url());
+
+    const web3 = new Web3(ulProvider);
+    web3.eth.defaultAccount = `0x${'00'.repeat(20)}`;
+
+    return {web3, ulProvider, services, deployedWallet};
+  };
+
   it('send transaction triggers wallet creationg', async () => {
-    const {web3, services} = await loadFixture(providerFixture);
+    const {web3, services} = await setup();
 
     expect(services.uiController.showOnboarding.get()).to.be.false;
 
@@ -20,5 +44,18 @@ describe('ULWeb3Provier', () => {
     await waitExpect(() => {
       return expect(services.uiController.showOnboarding.get()).to.be.true;
     });
+
+  });
+
+  it('can send a simple eth transfer', async () => {
+    const {web3, services, deployedWallet} = await setup();
+    services.walletService.connect(deployedWallet.asApplicationWallet);
+
+    const {transactionHash} = await web3.eth.sendTransaction({
+      to: Wallet.createRandom().address,
+      value: utils.parseEther('0.005').toString(),
+    });
+
+    expect(transactionHash).to.be.a('string');
   });
 });
