--- universal-login-web3/test/index.ts
@@ -1,5 +1,0 @@
-
-describe('ULWeb3Provider', () => {
-  describe('send', () => {
-  });
-});
